Rename do-telemetry to do-retry ahead of adding retries
Pure rename plus the four references, split out so the retry commit's diff
shows the file as a move rather than a delete/create pair -- the content
change is large enough that git's default 50% similarity threshold no longer
detects it.
